feat(code): Synchronized start of consensus when all validators/nodes connected (#190)
This is a quick hack to start consensus only when all nodes are connected.

## What does it do?

It is already programmed that the consensus actor receives gossip messages from the p2p layer. The implementation deferred all these messages until consensus was established. ( `handle_gossip_event` was only called if consensus Round is not `Nil`.) I restructured this part so two messages (specifically that a new peer has been connected or disconnected) are processed even if consensus has not been reached yet. I count the number of peers connected to the node while processing these messages. The peer IDs are stored in the State under `connected_peers`.

When a new peer connects, a check is made if the number of connected peers has reached the number of validators we expect on the network. If so, consensus is started.

This approach has multiple problems; hence, I consider it a hack.

* I think p2p messages shouldn't be processed in the Consensus actor. They should be processed in their own "Gossip" actor, and if necessary, a message should be sent to Consensus. This would require a bigger refactor, which I didn't want to get into until we have better (genesis) configuration options. However, now I rely on this to count the number of connected peers in Consensus.
* libp2p can more easily query the number of connected peers. This could be queried in the "Gossip" actor, and when enough have connected, a message could be sent to the Consensus actor. If the above refactor is made, I would rather use that method to start consensus.
* There will be other scenarios where we'll need to halt consensus, so maybe we should implement a start and a stop message and properly signal the Consensus actor.

---------

Co-authored-by: Romain Ruetschi <romain@informal.systems>