fix(code/discovery): Improve discovery performance (#483)
This PR addresses the performance issues of the current discovery implementation. Indeed, when spawning a local network of about 150 nodes, the current implementation leads to some APPLICATION_ERROR as the nodes are flooded with incoming dials and requests.

The new features are:

Requests now implement a retry mechanism similar to dials;
Both dials and requests are now rate-controlled with buffer channels.
For reference, with an M2 Pro and 32GB of RAM, this fixed version (with default config) spawns a (fully connected) local network of 150 nodes* in about 4.5 seconds without errors. Note that the speed can likely be optimized by tweaking some parameters.

*each node was provided with a single bootstrap node.

---

* fix: request retries, idle condition

* feat: dial/request buffer channels

* fix: applied @romac suggestions

* Revert code change based on my erroneous suggestion

---------

Co-authored-by: Romain Ruetschi <romain@informal.systems>