chore(code): Update libp2p to v0.55 (#772)
libp2p v0.55.0 comes with a bunch of new features and fixes to their GossipSub implementation, namely support for GossipSub v.1.2 which allows broadcasting IDONTWANT messages when publishing to avoid downloading data that is already available.

I have not opted-in to GossipSub v1.2 IDONTWANT in this PR, as it would be nice to benchmark the difference between having it enabled or not and play around with various thresholds before committing to it.

This release also includes a fix for an issue where an InsufficientPeers error could occur under certain conditions, despite having peers subscribed to a topic, which we are often seeing in tests.

Since libp2p now requires Rust 1.83, this PR also bumps our MSRV accordingly.

---

* chore(code): Update libp2p to v0.55.x

* Bump MSRV to Rust 1.83