fix(code/core-consensus): Only cancel propose timeout on a step change (#862)
* fix(code/core-consensus): Only cancel propose timeout on a step change

We were previously cancelling the propose timeout unconditionally after receiving a proposal,
but there is a case where we must not do so.

When we receive a proposal `Proposal(h, r, v, vr)` with `vr > -1` and without a polka for `vr`,
then we must stay in the propose step and wait for the timeout to expire so that
we either get more votes in the meantime that form a polka and we can prevote for `v` and
move to step prevote, or the timeout expires and we prevote `nil` and move to step prevote.

* Fix typo in doc comment