chore(CCHAIN275): Proposal keeper panics with batch/parallel sync and dynamic validator set (#1177)
Closes: #XXX

With sync ahead and batching we panic if validator set is dynamic. The
problem is that we determine the proposer of height `H` (the height of
the value in the response) based on the validator set for height `h`
(the current height when the response is received). We store this
proposer and when we reach `H` we verify that the proposer recorded is
what we expect for `H`.

https://github.com/circlefin/malachite/blob/cefb3d752a0cc57dc2b45b6c165e2330c2c35187/code/crates/core-driver/src/proposal_keeper.rs#L79-L90

The change proposed here is to buffer all responses for higher heights.
When the height is reached we process the response in the handler
`on_value_response()`

There are two flows resulting in `on_value_response()` being called:


  **Response coming from the network**
1.   `NetworkEvent::SyncResponse` (consensus.rs:484-507)
2.   `process_sync_response()` (consensus.rs:502) 
3.   `process_input()` with `SyncValueResponse` (consensus.rs:662-670)
4.   `malachitebft_core_consensus::process!()` (consensus.rs:314)
5.   `handle_input()` (handle.rs:40-68)
6.   `Input::SyncValueResponse` match (handle.rs:60)
7.   `on_value_response()` (sync.rs:6) 
8. `if response height is > current height` we buffer, otherwise we
process

**Later, when height advances on StartHeight Message**
1. `reset_and_start_height()` (start_height.rs:6)
2. `replay_pending_msgs()` (start_height.rs:64) 
3. `shift_and_take()` from Input Queue (start_height.rs:81)
4. `handle_input()` for each pending input (start_height.rs:91)
5. `Input::SyncValueResponse` match (handle.rs:60)
6. `on_value_response()` (sync.rs:6)
→ `if consensus_height < cert_height` re-buffer like above, else
→ process certificate if invalid send `Effect::InvalidValue` to the
engine (engine will notify Sync)
→ if valid send `Effect::SyncValue` to the engine (engine will notify
the application)

The proposer is set once in the engine on `Effect::SyncValue`, when the
response height matches the current height and the validator set is
known.


---------

Signed-off-by: Romain Ruetschi <github@romac.me>
Co-authored-by: Romain Ruetschi <romain.ruetschi@circle.com>