fix(driver): Produce `InvalidProposalAndPolkaPrevious` when receiving a polka certificate matching the POL round of a proposal with an invalid value
Test units that reveals the following scenario:

> Should we vote nil
[here](https://github.com/circlefin/malachite/blob/b205f4252f3064d9a74716056f63834ff33f2de9/code/crates/core-driver/src/mux.rs#L281)
if the certificate points to an invalid proposal (L32)? Also,
realistically this should never happen, right?

and the fix, which consist on producing a
`RoundInput::InvalidProposalAndPolkaPrevious` consensus state-machine
input when receiving a Polka certificate matching the POL round of a
proposal with an invalid value.

Notice that this is a very unlikely scenario, as this would mean that 2f
+ 1 processes have accepted the value in a previous round, while the
current process rejects it as an invalid value.

---

### PR author checklist

#### For all contributors

- [ ] Reference a GitHub issue
- [ ] Ensure the PR title follows the [conventional
commits][conv-commits] spec
- [ ] Add a release note in [`RELEASE_NOTES.md`](/RELEASE_NOTES.md) if
the change warrants it
- [ ] Add an entry in [`BREAKING_CHANGES.md`](/BREAKING_CHANGES.md) if
the change warrants it

#### For external contributors

- [ ] Maintainers of Malachite are [allowed to push changes to the
branch][gh-edit-branch]

[conv-commits]: https://www.conventionalcommits.org/en/v1.0.0/#summary
[gh-edit-branch]:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork#enabling-repository-maintainer-permissions-on-existing-pull-requests

---------

Co-authored-by: Romain Ruetschi <romain.ruetschi@circle.com>