fix(CCHAIN-658): Allow application to change its mind about validity (Invalid --> Valid)
Closes: #XXX

This addressed the case when the application changes its mind about the
validity of a value (e.g. because of a software upgrade).

If the validity of a value is not `Valid`, and used to be `Invalid`,
there are two code paths in `core-consensus` and the driver that won't
update the value in the corresponding keeper.
As a result, functions that hit those paths (e.g., `on_proposal`) will
then give the false impression that a value, with its validity info has
been stored.

This has hit an `assert!` in the devnet.

N.b.: We also removed some methods that prevented persisting
ProposedValue inputs (from the application) to the application when
duplicated.
This had consequences when the validity changed, so -- to be on the safe
side -- we are removing those optimizations and persisting every
ProposedValue including its validity to the WAL.

**Story:** https://circlepay.atlassian.net/browse/CCHAIN-658

---

- [ ] Reference a GitHub issue
- [ ] Ensure the PR title follows the [conventional
commits][conv-commits] spec
- [ ] Add a release note in [`RELEASE_NOTES.md`](/RELEASE_NOTES.md) if
the change warrants it
- [ ] Add an entry in [`BREAKING_CHANGES.md`](/BREAKING_CHANGES.md) if
the change warrants it

- [ ] Maintainers of Malachite are [allowed to push changes to the
branch][gh-edit-branch]

[conv-commits]: https://www.conventionalcommits.org/en/v1.0.0/#summary
[gh-edit-branch]:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork#enabling-repository-maintainer-permissions-on-existing-pull-requests

---------

Co-authored-by: Anca Zamfir <zamfiranca@gmail.com>
Co-authored-by: Daniel <daniel.cason@circle.com>
Co-authored-by: Romain Ruetschi <romain.ruetschi@circle.com>