chore: Remove initial height and validator set from consensus params (#1190)
Closes: #XXX

These parameters are a remnant from a very early time in Malachite's
lifespan where consensus would always start at the initial height with a
static validator set and move from height to height on its own.

These parameters have therefore not had any effect for a while, given
that it is now up to the app to tell consensus which height to start at
after receiving the `ConsensusReady` message.

As such, they can be removed without any loss of functionality.

---

### PR author checklist

#### For all contributors

- [ ] Reference a GitHub issue
- [ ] Ensure the PR title follows the [conventional
commits][conv-commits] spec
- [ ] Add a release note in [`RELEASE_NOTES.md`](/RELEASE_NOTES.md) if
the change warrants it
- [ ] Add an entry in [`BREAKING_CHANGES.md`](/BREAKING_CHANGES.md) if
the change warrants it

#### For external contributors

- [ ] Maintainers of Malachite are [allowed to push changes to the
branch][gh-edit-branch]

[conv-commits]: https://www.conventionalcommits.org/en/v1.0.0/#summary
[gh-edit-branch]:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/working-with-forks/allowing-changes-to-a-pull-request-branch-created-from-a-fork#enabling-repository-maintainer-permissions-on-existing-pull-requests